Подправил таблицы

+++ libs/ChatRole.py
@@ -7,7 +7,7 @@ class ChatRole(db.Model):
     __tablename__ = 'chat_roles'
 
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
-    chat_id = db.Column(db.Integer, db.ForeignKey('chats.id'), nullable=False)
+    chat_id = db.Column(db.BigInteger, db.ForeignKey('chats.id'), nullable=False)
     roles = db.Column(db.JSON, default=json.dumps({}))
 
     __table_args__ = (
